backend/requirement.txt
fastapi
uvicorn
langchain
langgraph
langchain-google-genai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
pytz